Smooth a numeric array with a Gaussian kernel

This function smoothes an array using a Gaussian kernel with a specified standard deviation.

gaussianSmooth(x, sigma)

Arguments

x
An object that can be coerced to an array, or for which a morph method exists.

sigma
A numeric vector giving the standard deviation of the kernel in each dimension. Can have lower dimensionality than the target array.

Value

A morphed array with the same dimensions as the original array.

Details

This implementation takes advantage of the separability of the Gaussian kernel for speed when working in multiple dimensions. It is therefore equivalent to, but much faster than, directly applying a multidimensional kernel.

See also

Author

Jon Clayden <code@clayden.org>